todo-backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djoser==2.3.3
django-cors-headers==4.7.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
gunicorn==21.2.0
dj-database-url==2.1.0 